+++ utils/visualize_fields/write_xdmf_box.py
@@ -14,10 +14,10 @@ non_uniform_grid = True
 #
 dtype_saves = np.dtype([                                                         \
                         ('file' , 'U100'), ('variable', 'U100'),                 \
-                        ('imin' , np.int), ('jmin' , np.int), ('kmin' , np.int), \
-                        ('imax' , np.int), ('jmax' , np.int), ('kmax' , np.int), \
-                        ('istep', np.int), ('jstep', np.int), ('kstep', np.int), \
-                        ('time', np.float), ('isave', np.int)                    \
+                        ('imin' , int), ('jmin' , int), ('kmin' , int), \
+                        ('imax' , int), ('jmax' , int), ('kmax' , int), \
+                        ('istep', int), ('jstep', int), ('kstep', int), \
+                        ('time', float), ('isave', int)                    \
                        ])
 geofiles  = "geometry_b_???.out"
 nblocks = np.size(glob.glob(geofiles))
